requirements.txt
python-telegram-bot==20.7
pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9 